+++ packages/db/prisma/migrations/20260814110417_add_event_image_asset_original_crop/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "EventImageAsset" ADD COLUMN     "crop" JSONB,
+ADD COLUMN     "original_url" TEXT;